band_server/requirements.txt
certifi==2022.9.24
grpcio==1.50.0
grpcio-tools==1.50.0
Pillow==9.3.0
pip==22.2.2
protobuf==4.21.9
setuptools==65.5.0
six==1.16.0
wheel==0.37.1
